[RCP-406] :sparkles: 회원가입 sqs 추가
